Added an API for bind shopfloors commands

+++ myems-api/core/command.py
@@ -267,6 +267,19 @@ class CommandItem:
                                    title='API.BAD_REQUEST',
                                    description='API.THERE_IS_RELATION_WITH_STORES')
 
+        # check relation with shopfloor
+        cursor.execute(" SELECT shopfloor_id "
+                       " FROM tbl_shopfloors_commands "
+                       " WHERE command_id = %s ",
+                       (id_,))
+        rows_shopfloors = cursor.fetchall()
+        if rows_shopfloors is not None and len(rows_shopfloors) > 0:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.THERE_IS_RELATION_WITH_SHOPFLOORS')
+
         # todo: check relation with points
 
         cursor.execute(" DELETE FROM tbl_commands WHERE id = %s ", (id_,))
